0.0.2 Base file writing
Added the ability to save the data currently available in program. Saves to txt file. The system should be able to support an arbitrary order of data when reading, as each set of info will have a string starting marker ("TRK;")(3 letters and semi colon) and a string ending marker ("E;"). This should also allow for versions to be forward compatible at the minimum needed sets of features, which should be similar to what was saved in Beta 2 versions.

2K lines produces a 28kb file, while 3.4.X produces a 262kb .sol file. Will need to compare these values when proper saving is established.