0.0.3 Visual and haxelib update
Updated to the latest versions of Haxe and OpenFL. With the updated openFL I can produce a more light weight import of .swf assets from the original beta 2. Currently Bosh's body is all that's visible.

To take advantage of this, you need a copy of flash professional or any software that can publish as a swf. Using the command 'openfl process name.swf', it will produce a directory ending with '.bundle', which can them be imported as a movieclip instead of a library.