0.0.3 FIle browser base
Made a UI to show files in the saves folder. Originally I wanted it to have a file dialog pop up and you simply browse for your track, but that wasn't works. Saves need to be placed in the saves folder in order to show up on the list. http://imgur.com/a/ks5fv