Reworking toolbar
Prepwork to make toolbar easy to scale without losing quality. This is a bit of an unorganized commit because I am tired, but anyone outside viewing should understand what's going on if they compare the code to a few other visual items, such as the start point, or bosh's body.