Massive overhaul of toolbar system
Way too lazy to individually commit each detail change. Long story short, instead of using .png's for icons, it's now being upgraded to a flash asset. This will allow in the future for scaleable UI without losing quality when scaled. A setting will soon be added for this.

Not all icons have been added, but this commit is to show the system works, it's agnostic, and allows for far easier additions to newer icons that might need be added.